examples/rag-llamaindex-python/requirements.txt
blindfold-sdk>=1.3.0
llama-index>=0.12.0
llama-index-llms-openai>=0.4.0
llama-index-embeddings-openai>=0.3.0
python-dotenv>=1.0.0
